assign channels to associated parties

+++ app/Main.hs
@@ -3,7 +3,6 @@ module Main (main) where
 import Lib
 import Parser
 import Connect
-import Data.Int
 import System.Process
 import System.IO
 import System.Environment
@@ -22,7 +21,7 @@ main = do
   return ()
 
 
-createProcessLoop :: Int16 -> Int16 -> Int16 -> IO ()
+createProcessLoop :: Int -> Int -> Int -> IO ()
 createProcessLoop n m pid
   | n == 0 || pid > 0 = return ()
   | otherwise = do
